+++ lib/parser.dart
@@ -2,26 +2,24 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library csslib.parser;
-
 import 'dart:math' as math;
 
 import 'package:source_span/source_span.dart';
 
 import 'src/messages.dart';
-import 'src/options.dart';
+import 'src/preprocessor_options.dart';
 import 'visitor.dart';
 
-export 'src/messages.dart' show Message;
-export 'src/options.dart';
+export 'src/messages.dart' show Message, MessageLevel;
+export 'src/preprocessor_options.dart';
 
 part 'src/analyzer.dart';
 part 'src/polyfill.dart';
 part 'src/property.dart';
 part 'src/token.dart';
-part 'src/tokenizer_base.dart';
+part 'src/token_kind.dart';
 part 'src/tokenizer.dart';
-part 'src/tokenkind.dart';
+part 'src/tokenizer_base.dart';
 
 enum ClauseType {
   none,