+++ lib/src/tokenizer.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of csslib.parser;
+part of '../parser.dart';
 
 class Tokenizer extends TokenizerBase {
   /// U+ prefix for unicode characters.