+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.15.0
+
+- **BREAKING**
+  - Removed `css` executable from `bin` directory.
+  - Removed the deprecated `css.dart` library.
+  - `Message.level` is now of type `MessageLevel` defined in this package.
+- Removed dependencies on `package:args` and `package:logging`.
+- Require Dart SDK `>=2.1.0`.
+
 ## 0.14.6
 
 * Removed whitespace between comma-delimited expressions in compact output.